[f] end pharse1

+++ TaLa/Model/Models/Player+CoreDataProperties.swift
@@ -18,7 +18,7 @@ extension Player {
     }
 
     @NSManaged public var name: String?
-    @NSManaged public var score: Int64
+    @NSManaged public var score: Int16
     @NSManaged public var id: Int16
     @NSManaged public var image: String?
     @NSManaged public var game: Game?
